ricoh-theta-module-master/requirements.txt
# core
cython
setuptools
nose

# math/tensors
numpy
scipy

# plot
matplotlib

# io
argparse
pyyaml

# extra
imutils
